Simple Command Framework
This is very similar to the original PhaseBot framework, but there
wasn't anything insanely wrong with it. An annotation system could be
cool, but that adds slowness and it isn't needed.

A simple string repsonse command would be neat, as in a user runs the
command:

    /text1

and in a file named `text.txt`:

    Wow, this is some text!

The bot will send a chat message containing that message.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>